% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdown-package.R
\docType{package}
\name{bookdown-package}
\alias{bookdown}
\alias{bookdown-package}
\title{Authoring Books and Technical Documents with R Markdown}
\description{
A open-source (GPL-3) R package to facilitate writing books and long-form
articles/reports with R Markdown.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/bookdown}
  \item \url{https://pkgs.rstudio.com/bookdown/}
  \item Report bugs at \url{https://github.com/rstudio/bookdown/issues}
}

}
\author{
\strong{Maintainer}: Yihui Xie \email{xie@yihui.name} (\href{https://orcid.org/0000-0003-0645-5666}{ORCID})

Other contributors:
\itemize{
  \item JJ Allaire [contributor]
  \item Albert Kim [contributor]
  \item Alessandro Samuel-Rosa [contributor]
  \item Andrzej Oles [contributor]
  \item Atsushi Yasumoto (\href{https://orcid.org/0000-0002-8335-495X}{ORCID}) [contributor]
  \item Aust Frederik (\href{https://orcid.org/0000-0003-4900-788X}{ORCID}) [contributor]
  \item Bastiaan Quast [contributor]
  \item Ben Marwick [contributor]
  \item Chester Ismay [contributor]
  \item Christophe Dervieux [contributor]
  \item Clifton Franklund [contributor]
  \item Daniel Emaasit [contributor]
  \item David Shuman [contributor]
  \item Dean Attali [contributor]
  \item Drew Tyre [contributor]
  \item Ellis Valentiner [contributor]
  \item Frans van Dunne [contributor]
  \item Hadley Wickham [contributor]
  \item Jeff Allen [contributor]
  \item Jennifer Bryan [contributor]
  \item Jonathan McPhers [contributor]
  \item JooYoung Seo (\href{https://orcid.org/0000-0002-4064-6012}{ORCID}) [contributor]
  \item Joyce Robbins [contributor]
  \item Junwen Huang [contributor]
  \item Kevin Cheung [contributor]
  \item Kevin Ushey [contributor]
  \item Kim Seonghyun [contributor]
  \item Kirill Muller [contributor]
  \item Luciano Selzer [contributor]
  \item Matthew Lincoln [contributor]
  \item Maximilian Held [contributor]
  \item Michael Sachs [contributor]
  \item Michal Bojanowski [contributor]
  \item Nathan Werth [contributor]
  \item Noam Ross [contributor]
  \item Peter Hickey [contributor]
  \item Pedro Rafael D. Marinho (\href{https://orcid.org/0000-0003-1591-8300}{ORCID}) [contributor]
  \item Romain Lesur (\href{https://orcid.org/0000-0002-0721-5595}{ORCID}) [contributor]
  \item Sahir Bhatnagar [contributor]
  \item Shir Dekel (\href{https://orcid.org/0000-0003-1773-2446}{ORCID}) [contributor]
  \item Steve Simpson [contributor]
  \item Thierry Onkelinx (\href{https://orcid.org/0000-0001-8804-4216}{ORCID}) [contributor]
  \item Vincent Fulco [contributor]
  \item Yixuan Qiu [contributor]
  \item Zhuoer Dong [contributor]
  \item  RStudio, PBC [copyright holder]
  \item Bartek Szopka (The jQuery Highlight plugin) [contributor]
  \item Zeno Rocha (clipboard.js library) [copyright holder]
  \item  jQuery Foundation (jQuery library) [copyright holder]
  \item  jQuery contributors (jQuery library; authors listed in inst/resources/AUTHORS) [contributor, copyright holder]
  \item  MathQuill contributors (The MathQuill library; authors listed in inst/resources/AUTHORS) [contributor]
  \item  FriendCode Inc (The gitbook style, with modifications) [copyright holder, contributor]
}

}
\keyword{internal}
