% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{create_book}
\alias{create_book}
\alias{create_gitbook}
\alias{create_bs4_book}
\title{Create a bookdown projet}
\usage{
create_gitbook(path)

create_bs4_book(path)
}
\arguments{
\item{path}{Path where to create the bookdown project. In RStudio IDE,
if \pkg{rstudioapi} package available, a RStudio project will be created.}
}
\description{
Create a basic bookdown project for an HTML book. This will create a folder
with file structure for a bookdown project, and example files with
information on how to start. This template book can be built directly
}
\details{
The book created will use one of the HTML book format available:
\itemize{
\item Use \code{create_gitbook()} to use \code{bookdown::gitbook()} output format
\item Use \code{create_bs4_book()} to use a \code{bookdown::bs4èbook()} output format
}
}
