% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{number_caption}
\alias{number_caption}
\title{Customize numbering of figures and tables}
\usage{
number_caption(
  n,
  labels,
  format = function(label, i) sprintf("\%s \%s: ", label, i),
  format2 = function(labels, i) sprintf("\%s \%s: ", labels[1], i)
)
}
\arguments{
\item{n}{A vector of breakpoints (sorted increasingly). It can be either
numeric or character. If a breakpoint contains a fraction that ends with 0,
it must be written in character, e.g., \code{"2.10"} (otherwise it will be
treated as \code{2.1}).}

\item{labels}{A vector of caption labels of \code{length(n) + 1}. The first
label is used when the current caption number \code{i} is smaller than or
equal to \code{n[1]}, and the second label is used when \code{i <= n[2]},
etc.}

\item{format}{A function that accepts the current label and number, and
returns a character string as the caption prefix, e.g., \samp{Figure 1:}.}

\item{format2}{A function to generate the caption prefix when the caption
number is not numeric, e.g., when the caption is in an appendix and the
number may be \samp{A.1}.}
}
\value{
A function that can be passed to the \code{fig} or \code{tab} fields
  in \file{_bookdown.yml}.
}
\description{
Figure/table numbering can be customized via functions (see
\url{https://bookdown.org/yihui/bookdown/internationalization.html}). This
function provides a way to customize the numbering of captions according to
the ranges of caption numbers. For example, a different caption label can be
used if the caption number is greater than a value.
}
\examples{
f = bookdown::number_caption(8, c("Table", "Supplementary table"))
f(1)
f(8)
f(8.5)
f(9.1)
}
