% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish_book}
\alias{publish_book}
\title{Publish a book to a Posit Connect server}
\usage{
publish_book(name = NULL, account = NULL, server = "connect.posit.cloud", ...)
}
\arguments{
\item{name}{Name of the book (this will be used in the URL path of the
published book). Defaults to the \code{book_filename} in
\code{_bookdown.yml} if not specified.}

\item{account}{Account name to publish to. Will default to any previously
published to account or any single account already associated with
\code{server}.}

\item{server}{Server to publish to (by default connect.posit.cloud, but any
Posit Connect server can be published to).}

\item{...}{Other arguments to be passed to [rsconnect::deploySite()].}
}
\description{
Publish a book to a Connect Server. By default, you should render the book
locally before publishing.
}
\note{
Previously the default server was bookdown.org, which will be sunset.
  You are no longer recommended to publish to bookdown.org.
}
