% Please edit documentation in R/site.R
\name{site_generators}
\alias{gitbook_site}
\alias{html_book_site}
\alias{site_generators}
\alias{tufte_html_book_site}
\title{R Markdown site generators for bookdown}
\usage{
gitbook_site(input, ...)

html_book_site(input, ...)

tufte_html_book_site(input, ...)
}
\arguments{
\item{input}{Website directory (or the name of a file within the directory)}

\item{...}{Currently unused}
}
\description{
Implementation of custom R Markdown site generators for bookdown.
}
