% Please edit documentation in R/site.R
\name{bookdown_site}
\alias{bookdown_site}
\title{R Markdown site generator for bookdown}
\usage{
bookdown_site(input, ...)
}
\arguments{
\item{input}{Website directory (or the name of a file within the directory)}

\item{...}{Currently unused}
}
\description{
Implementation of custom R Markdown site generator for bookdown.
}
