% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs4_book.R
\name{bs4_book}
\alias{bs4_book}
\alias{bs4_book_theme}
\title{HTML book built with bootstrap4.}
\usage{
bs4_book(
  theme = bs4_book_theme(),
  ...,
  lib_dir = "libs",
  pandoc_args = NULL,
  extra_dependencies = NULL
)

bs4_book_theme(...)
}
\arguments{
\item{theme}{A named list or [bootstraplib::bs_theme()] object.
The default, `bs4_book_theme()` resets the base font size to 1rem, to
make reading easier. A named list will be passed to `bs4_book_theme()`,
making it possible to specify theme settings in the yaml metadata.}

\item{lib_dir, pandoc_args, extra_dependencies, ...}{Passed on to
[rmarkdown::html_document()].}
}
\description{
This output format is designed to give a clean reading experience.
Main features:

* A per section-search engine that helps you quickly find what you're
  looking for.

* A sidebar that includes a within-chapter table of contents that
  dynamically updates so it's easy to keep track of where you are
  within the chapter.

* Thoughtful typography to make the contents as easy as possible to read,
  regardless of the size of your device. A sticky header gets out of your
  way when reading, but is easily accessible if you need it.

* In-line footnotes.

* R syntax highlighting and autolinking by
  [downlit](http://downlit.r-lib.org/) is paired with a accessible
  theme designed by Alison Hill.

* The ability to customise colours and fonts through
  [bootstraplib](https://rstudio.github.io/bootstraplib)
}
\section{Limitations}{


* `bs_book4()` is designed specifically for books that use one chapter per
  page.
}

