% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_fenced_div}
\alias{convert_to_fenced_div}
\title{Convert theorem and proof engine to new fenced div syntax}
\usage{
convert_to_fenced_div(file, modify = FALSE)
}
\arguments{
\item{file}{Path to a file to modify.}

\item{modify}{Set to \code{TRUE} to overwrite the file with the Theorems and
Proofs environments changed.}
}
\value{
\code{TRUE} or \code{FALSE} invisibly if the file has been modified
  or not.
}
\description{
This function is helper to quickly transition from the old syntax to the knew one.
It will parse a R Markdown file to replace any chunk using one of
the theorems or proofs engine to use the new syntax using Pandoc's fenced div.
}
\details{
In the chunk header, it will replace backticks with colons, use the engine as
class by adding a dot, use the explicit or implicit chunk label as id by
adding a dash and keep the other attributes as is. We strongly advice to
check the changes and adjust if necessary.
}
\section{Warning}{


  \strong{Use with care !}. This function overwrite the file if an engine
  needs to me modified. We strongly advice to use on files that are under
  version control or to create backup copy.

  You should check the modifications, especially is you used theorem or proof
  engine in a non standard way.
}

\references{
\itemize{
  \item \href{https://bookdown.org/yihui/bookdown/markdown-extensions-by-bookdown.html#theorems}{About Theorems and proofs}
  \item \href{https://bookdown.org/yihui/rmarkdown-cookbook/custom-blocks.html}{About custom blocks}
}
}
