% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish_book}
\alias{publish_book}
\title{Publish a book to the web}
\usage{
publish_book(name = NULL, account = NULL, server = NULL, render = c("none", "local", 
    "server"))
}
\arguments{
\item{name}{Name of the book (this will be used in the URL path of the
published book). Defaults to the \code{book_filename} in
\code{_bookdown.yml} if not specified.}

\item{account}{Account name to publish to. Will default to any previously
published to account or any single account already associated with
\code{server}.}

\item{server}{Server to publish to (by default beta.rstudioconnect.com but
any RStudio Connect server can be published to).}

\item{render}{Rendering behavior for site: "none" to upload a
static version of the current contents of the site directory;
"local" to render the site locally then upload it; "server" to
render the site on the server. Note that for "none" and "local"
R scripts (.R) and markdown documents (.Rmd and .md) will not be
uploaded to the server.}
}
\description{
Publish a book to the web. Note that you should be sure to render all
versions of the book before publishing, unless you have specified
\code{render = TRUE}.
}
